fix:#04 no draw is reason by fragment texname
