+++ src/03_09_effect1/index.html
@@ -23,8 +23,8 @@
       <div id="page-container">
         <div
           id="div1"
-          data-webgl="01_sticky-plane"
-          data-tex-1="/img/output5.jpg"
+          data-webgl="01sticky-plane"
+          data-tex-1="/img/output4.jpg"
           data-tex-2="/img/output6.jpg"
         ></div>
         <!-- <div
