+++ src/03_09_effect1/scripts/glsl/01sticky-plane/fragment.glsl
@@ -3,12 +3,12 @@ precision mediump float;
 varying vec2 vUv;
 varying float vDelay;
 
-uniform sampler2D uTex;
+uniform sampler2D tex1;
 uniform float uProgress;
 
 void main() {
 
-  vec4 tex = texture(uTex, vUv);
+  vec4 tex = texture(tex1, vUv);
   
   // gl_FragColor = vec4(vDelay, 0., 0., 1.);
   gl_FragColor = tex;
